# Remove redundant regexp