The same processing is possible without using regular expressions.

−               | before   | after   | result
--               | --       | --      | --
parse            | 61.006   | 61.975  | 1.016x
render           | 197.095  | 197.103 | -
parse & render   | 44.846   | 45.180  | 1.008x

### Environment
- MacBook Pro (14 inch, 2021)
- macOS 12.3 beta
- Apple M1 Max
- Ruby 3.0.3

### Before
```
$ ruby benchmark.rb

Running benchmark for 10 seconds (with 5 seconds warmup).

Warming up --------------------------------------
              parse:     6.000  i/100ms
             render:    19.000  i/100ms
     parse & render:     4.000  i/100ms
Calculating -------------------------------------
              parse:     61.006  (± 0.0%) i/s -    612.000  in  10.032337s
             render:    197.095  (± 1.5%) i/s -      1.976k in  10.027538s
     parse & render:     44.846  (± 2.2%) i/s -    452.000  in  10.080806s
```

### After
```
$ ruby benchmark.rb

Running benchmark for 10 seconds (with 5 seconds warmup).

Warming up --------------------------------------
              parse:     6.000  i/100ms
             render:    19.000  i/100ms
     parse & render:     4.000  i/100ms
Calculating -------------------------------------
              parse:     61.975  (± 0.0%) i/s -    624.000  in  10.069036s
             render:    197.103  (± 1.0%) i/s -      1.976k in  10.026615s
     parse & render:     45.180  (± 0.0%) i/s -    452.000  in  10.005335s
```