@Watson1978 2022-03-16
> I don't think token.strip is a good idea. If you revert this change, rest looks good to me.

OK, reverted at https://github.com/Shopify/liquid/pull/1540/commits/22568080b1cb6f19969d55d68dc39b5c04a37b17

However, I think it should use `String#match?` instead because it doesn't use backreference of `String#=~`. 
See https://github.com/Shopify/liquid/pull/1540/commits/fad58ef436bb9a516ea8364d9398cb6f6876d784